add : user authenticate

+++ server/controllers/topics.controllers.js
@@ -14,23 +14,63 @@ const getTopic = async (req, res) => {
   return res.status(200).send(topic);
 };
 
+// 글 작성
+const newTopic = async (req, res) => {
+  const { title, content, img_url } = req.body;
+  let info = {
+    title: title,
+    content: content,
+    img_url: img_url,
+    user_id: req.userId,
+  };
+  const topic = await Post.create(info);
+  res.status(200).send(topic);
+
+  const fs = require("fs");
+  var DEPLOYED_ABI = JSON.parse(
+    fs.readFileSync("../contract/20deployedABI", "utf8")
+  );
+  var DEPLOYED_ADDRESS = fs
+    .readFileSync("../contract/20deployedAddress", "utf8")
+    .replace(/\n|\r/g, "");
+  const Contract20 = new web3.eth.Contract(DEPLOYED_ABI, DEPLOYED_ADDRESS);
+};
+
 const updateTopic = async (req, res) => {
   let id = req.params.pid;
-  const topic = await Post.update(req.body, { where: { id } });
-  return res.status(200).send(topic);
+  const topic = await Post.findOne({ where: { id } });
+
+  if (topic.user_id === req.userId) {
+    const { title, content, img_url } = req.body;
+    let info = {
+      title: title,
+      content: content,
+      img_url: img_url,
+    };
+    await Post.update(info, { where: { id } });
+    const topic = await Post.findOne({ where: { id } });
+    return await res.status(200).send(topic);
+  } else {
+    return res.status(403).send({ message: "해당 게시물의 소유주가 아닙니다" });
+  }
 };
 
 const deleteTopic = async (req, res) => {
   let id = req.params.pid;
-  const topic = await Post.destroy({ where: { id } });
-  console.log(topic);
-  return res.status(200).send("찾을 수 없는 게시물입니다");
+  const topic = await Post.findOne({ where: { id } });
+  if (topic.user_id === req.userId) {
+    await Post.destroy({ where: { id } });
+    return res.status(200).send({ message: "삭제되었습니다" });
+  } else {
+    return res.status(403).send({ message: "해당 게시물의 소유주가 아닙니다" });
+  }
 };
 
 module.exports = {
   getTopic,
   updateTopic,
   deleteTopic,
+  newTopic,
 };
 
 /*
